+++ src/impl/gameplay.c
@@ -207,8 +207,8 @@ bool draw_rack_from_bag(const Game *game, const int player_index,
   const uint16_t dist_size = rack_get_dist_size(player_rack);
   rack_copy(player_rack, rack_to_draw);
   for (int i = 0; i < dist_size; i++) {
-    const int rack_number_of_letter = rack_get_letter(player_rack, i);
-    for (int j = 0; j < rack_number_of_letter; j++) {
+    const int8_t rack_number_of_letter = rack_get_letter(player_rack, i);
+    for (int8_t j = 0; j < rack_number_of_letter; j++) {
       if (!bag_draw_letter(bag, i, player_draw_index)) {
         return false;
       }
@@ -224,8 +224,8 @@ void draw_leave_from_bag(Bag *bag, int player_draw_index, Rack *rack_to_update,
                          const Rack *rack_to_draw) {
   const uint16_t dist_size = rack_get_dist_size(rack_to_draw);
   for (int i = 0; i < dist_size; i++) {
-    const int rack_number_of_letter = rack_get_letter(rack_to_draw, i);
-    for (int j = 0; j < rack_number_of_letter; j++) {
+    const int8_t rack_number_of_letter = rack_get_letter(rack_to_draw, i);
+    for (int8_t j = 0; j < rack_number_of_letter; j++) {
       if (!bag_draw_letter(bag, i, player_draw_index)) {
         continue;
       }
